Renamed package logging.slf4j to integration
